прикрутил loff статус. Сброс незавершенных команд с компа по таймауту.

+++ ads1292.h
@@ -31,7 +31,7 @@ void AFE_Cmd(uchar cmd) ;
 void AFE_Write_Reg(uchar addr, uchar numOfBytes, const uchar* values);
 void AFE_Read_Reg(uchar addr, uchar numOfBytes, uchar* regBuf);
 void AFE_Read_Data(long* result);
-uchar AFE_getLoffStatus();
+uint AFE_getLoffStatus();
 void AFE_StartRecording();
 void AFE_StopRecording();
